moved tests to correct folder
